Revert "Fix Xcode 5 warnings and errors"
This reverts commit 2fc2f1fbe27a8177db0b2d6da70bdced824b5356.

+++ Source/CDRDefaultReporter.m
@@ -166,10 +166,10 @@
 
 - (void)printNestedFullTextForExample:(CDRExample *)example stateToken:(NSString *)token {
     static NSMutableArray *previousBranch = nil;
-    NSUInteger previousBranchLength = previousBranch.count;
+    int previousBranchLength = previousBranch.count;
 
     NSMutableArray *exampleBranch = [example fullTextInPieces];
-    NSUInteger exampleBranchLength = exampleBranch.count;
+    int exampleBranchLength = exampleBranch.count;
 
     BOOL onPreviousBranch = YES;
